Initial implementation of Function variations (#1)
Implementation of a simple Python function that is responsible for taking the provided lookup key, checking the Redis cache, looking up in Cosmos DB if not found (and adding to Redis cache), and returning the details. The lookup key is the id property of the Volcano Data, and the details will be the full document found in Cosmos DB for that id.